chore(BridgeGame, BridgeMaker): move BridgeGame and BridgeMaker into bridge.domain package
